ArchiveContent: use fakeroot to extract the archive
This allows extracting a root filesystem archive with files
that usually could not be created by a non-root user
(i.e. device nodes or files owned by root).

The attributes are stored by fakeroot and later used again,
when the files are added to a partition (e.g. using Ext4Content or Fat32Content)